wasm: use a single alien thread for compilation
Using a 1:1 alien thread pool is wasteful in terms of memory (I believe
by default threads have ~8 MB stacks in linux?), so only use a single
thread. Possibly in the future we can temporarily spin up more threads
if we need to compile many modules (like on startup for instance).

Also we need to unblock wasmtime's signals on the reactor threads now
that we're running wasm directly on the reactor thread.

Signed-off-by: Tyler Rockwood <rockwood@redpanda.com>