schema_registry: Split project_ids
A common use of `POST /subjects/{subject}/versions` is to get the
schema ID of an existing subject-version.

This case can be optimised by not taking the write_sem unless a new
schema or version needs to be written to the topic.

Split out `get_schema_version` from `project_ids` to facilitate that.

Signed-off-by: Ben Pope <ben@redpanda.com>