r/recovery_stm: check if node is a leader before sending request
Validation in `suppress_heartbeats_guard` checks if a node is a leader
when heartbeat is suppressed. The assertion is triggered if node
suppressing heartbeat is no longer a leader. Added a check immediately
preceding heartbeat suppression in `recovery_stm`. Recovery STM works
outside of the raft mutex and it may be the case that leadership was
lost between the fiber scheduling points.

Fixes: #14066

Signed-off-by: Michal Maslanka <michal@redpanda.com>