s3_client: Improve error handling
Currently, the S3 client interprets any response without body as
SlowDown response but it should only do this if S3 replied with 503.

This commits changes this so in case of error 503 the error will be
retriable and in case of any other error it will be  a normal error (the
error is represented as an internal error in this case).

(cherry picked from commit 8eed9d8088c3da3c8dc7c56cfe9bea7273fc6f72)