wasm: minimal poll_oneoff for golang
Implement poll_oneoff for sleeping. This allows for "sleeping" which
upstream golang assumes is present for it to work. Only sleeping is
required, which we implement via a noop that does nothing, as the clock
is fixed for each record that is being processed.

Upstream golang uses this function for usleep here:
https://github.com/golang/go/blob/7e1713e93c626f7026d1de09258f03748c726e79/src/runtime/os_wasip1.go#L165-L181

Signed-off-by: Tyler Rockwood <rockwood@redpanda.com>