admin: report cloud_log_segment_count
The metric is reported to RPK and used in `rpk topic describe-storage`.
While it's not totally accurate, on account of missing out on spillover
segments, we should still expose it.

I considered changing the field in RPK but this seems like the long-term
field we want, and we should just eventually improve its correctness in
Redpanda.

Fixes #14101