ducktape: Ignore bad log lines within a time window
Usually a test will ignore bad log strings throughout the test run. However
there are instances where we know exactly when the bad log lines are
acceptable but not elsewhere.

This change adds a bad log line matcher which is initialized with the
expression to match but no timestamps. Later the test can use the new
decorator added here to mark the time ranges where the expression is
acceptable.

When searching for bad log lines to ignore, both the expression and the
datetime range must match the log entry for it to be allowed.