archival: skip reuploads of identical size
PR #9404 partially fixed the issue of compacted reupoads of the same
size leading to gaps in the cloud log. However, the fix assumed the
issue can only happen when one segment is selected for reupload. That's
not the case. For example:

Node 1 segment layout: [0-2][3-3]
Node 2 segment layout: [0-3]

Node 1 does the initial upload of the offset range [0-3].
This is covered by just one segment. Then, Node 2 gains leadership.
It decides to do a compacted re-upload for the [0-3] range. This is
covered by two segments now. If the keys in the [0-3] range are unique,
the reupload will have the same size, but our fix doesn't kick in since
the reupload contains multiple segments.

PR #9404 was followed up by #9597 which added an additional defence
layer for this in the manifest. While the issue doen't lead to gaps in
the log, we still reupload a segment for no reason and replicate
commands which fail to apply.