metrics: Set aggregate_metrics.needs_restart=no
This is a pain for cloudv2 where we generally run with
`aggregate_metrics=true`.

Very often for debugging you want the unaggregated metrics. Having to
restart the cluster just to change this is quite interruptive.

This patch changes the flag itself. Subsequent patches will add support
to all the places where the flag is used.