s3_client: Check if response contains error
S3 can reply with error even if it's replying using status 200. Most of
our parsing is done independent of status code (we're trying to check
error fields every time). But in case of delete objects reply we're not
trying to parse error fields if the status is 200. This commit fixes
this and before any attempt at parsing the response is done we're
checking if the 'Error.Code' field is present.