tests: retry on timeout in partition balancer test
In partition balancer maintenance mode test one of the nodes is made
unavailable. In some cases the node may be suspended while being a
`redpanda/controller` partition leader. In this case changing
configuration may result in timeout, which is perfectly fine. We should
allow Admin client to retry in this case.

Fixes: #14047

Signed-off-by: Michal Maslanka <michal@redpanda.com>