ducktape/cloud_storage: Wait for at least two spillovers
When creating a manifest for reset we drop the first spillover and then
use the rest of the spillovers to craft a new manifest. If only one
spillover has been uploaded, then after dropping the first entry nothing
is left to craft the manifest.

This change makes the test wait until more than one manifest is spilled
over.