cloud_storage: add last scrubbed offset to manifest
This commit introduces a new field to the partition manifest:
`_last_scrubbed_offset`. It represents the offset at which the last
scrub stopped. Note that this offset is not linear. The scrubber
processes manifests from high offsets to low offsets, but the segments
within a manifest are processed from low to high.