transform/manager: track processor state
Add processor state to an entry in the processor table, we need to track
if a transform is errored or if we're starting the transform and is just
inactive as it's starting.

Additionally, we track the node that the transform_manager is running on
for reports (to be used in the next commit).

Signed-off-by: Tyler Rockwood <rockwood@redpanda.com>