storage: evict data from compactible topics under disk pressure
Currently, under disk pressure, we evict local data that was uploaded to
the cloud storage only for topics with `cleanup.policy` equal to
`delete` or `delete,compact`.

Due to an oversight, data for topics with the policy set to `compact` is
never evicted. This commit ensures consistent eviction behavior for all
cleanup policies.

Since Redpanda doesn't run compaction for data that is present only in
the cloud storage, this change introduces a potential regression in
compaction quality. This choice is considered better than running out of
disk space.