archival: don't stop the upload loop on sync fails
Previously, we'd bail out of the upload loop if the initial sync call
failed. The upload loop wouldn't restart until forced by re-setting the
remote write topic config or restarting the node or changing leadership.
This leads to the disk filling up since the collectable offset isn't
advancing due to the lack of uploads.

The fix is to swap the `co_return` with `continue` to allow for retries.
I've also added some error logging for when the loops quit unexpectedly.