config: helper for mutating node config globally
There isn't a compelling reason for modifying the config on a single
shard. There is an opportunity for bugs to creep in when used
incorrectly. This commit proposes to expose node local config as
readonly (const return type) and introduces a helper function that will
modify the configs on all shards.