rptest: allow for skipping of bucket deletion
Some scale tests produce huge buckets and it's not feasible to delete
them at the end of the test. This commit introduces a new flag to
SISettings, which allows the test to skip the bucket deletion. If said
flag is used, the test will apply a lifecycle policy to the bucket which
deletes its contents after a minimum of one day.