archival: add full scrubbing interval
If a scrub has not reached the end of the log, we call that a partial
scrub, otherwise it's a full scrub. For integrating scrubbing into our
ducktape testing suite, it will be useful to make the full scrub delay
larger. However, in production I expect both configs to have the same
value generally.

This commit teaches the `scrubber_scheduller` how to juggle the two
timestamps, introduces the new cluster configuration and threads it into
the rest of the `archival` code.