cluster/kv: track kv limit using bytes
Instead of reasoning about the amount of memory that the offset tracker
for wasm could use as number of keys constant, we instead compute the
max number of keys from the number of entries that we're OK with
storing.

Instead of a blanket 10K keys we can keep in the distributed_kv_stm, we
now default to 1MB worth of data can be stored. To keep the
implementation fast and simple, we assert that the memory usage for
each object is fixed at compile time.

In our specific case, for offset metadata we now can store ~20k
transform offsets instead of 10k.

If we ever need key value pairs with dynamic memory usage, we can
evolve the implementation.

Signed-off-by: Tyler Rockwood <rockwood@redpanda.com>