rpk: always use one newline in section commands
Sections _always_ added a newline at the end of them. Some commands
_also_ added a newline at the end inside a section.

This removes out.Header in favor of out.Sections, which is hopefully a
bit more foolproof to use so we do not mistakenly add newlines in the
wrong place. out.Sections only print newlines between sections. If we
try to add a section that was not registered, nothing happens -- this
allows the "conditional add" that we currently use in some areas.
Although this is a liiitttllee bit uglier, it is hopefully a bit more
understandable to use and more foolproof, rather than counting sections
ahead of time and having two or three booleans dictating who knows what
at the callsite.

The group describe code has been cleaned up such that each chunk that
prints text is only responsible for its own text, not the newline. The
newline is added by the caller and only if there is an extra block of
text that will follow the newline.