storage: don't force put() to take ownership of key
The full fidelity index will store a copy of the key, so taking it by
value is nice because the value can be moved into the index structure.

However for the hash-based version of the index the key is not stored.
In this case taking by reference is not an optimization to be concerned
about, but the linter will complain that we aren't actually moving or
copying the key.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>