storage: futurize offset map interface
The hash-based implementation of the map contains linear probing
technique that may possibly result in a reactor stall. We may find that
we don't want to pay the performance penality of a futurized interface
and that the linear probing is effectively never at risk of producing a
reactor stall, but we can back this out later if we have performance
issues.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>