cloud_storage: keep missing archive segments
Previously, we would remove any missing segment anomaly which did not
line up with a segment in the STM manifest. This is fine for segments
that have gone missing from the STM manifest, but not for segments
missing for the archive. Since we don't perform any form of reuploads
in the archive section, we should always keep these missing segments if
they're above the start offset of the archive.