storage: allow put on key collision even if map full
Updates the behavior of the key_offset_map to allow entries to be put
even if it's already full, provided it doesn't require adding new keys.

This makes certain aspects of testing easier (e.g. we can say that a map
with a limit of 10 keys will index all segments with a key space
containing 10 keys, instead of stopping immediately after we see the
10th key).