storage: allow keeping the last offset when compacting
In sliding window compaction, it will be possible that the entire
segment's data will be removed. To maintain that we still have data in
each segment, this allows passing the last segment offset to the
segment reducer, allowing it to tell whether it's empty by the end of
its reduce and force keeping a record accordingly.