wasm: add executable memory metrics
This technically isn't "all" executable in terms of `mprotect`,
as some of the JIT'ed code is read only data segments. This is
effectively how much memory wasmtime has allocated via `mmap` (with the
config settings we use).

This is important to have a metric for because the user needs to ensure each
node has enough memory for this data outside of what seastar has
allocated, since this data is allocated outside of seastar.

Signed-off-by: Tyler Rockwood <rockwood@redpanda.com>