transform/rpc/client: don't throw errc
This prevents us from returning futures that have exceptions which are
error codes. As far as I know we only throw exceptions (and return
exceptional futures which hold a subclass of `std::exception`), and this
feels sneaky in case someone tries to catch these via `catch (const std::exception&)`.

This does nullify the retry behavior but that will be addressed in the
next commit.

Signed-off-by: Tyler Rockwood <rockwood@redpanda.com>