archival: use rtc nodes owned by housekeeping jobs
Each housekeeping job will now run under its own rtc node. This allows
jobs to shut down correctly. Previously, a job could try to abort, but
the retry chain node that it was using would not listen on its abort
source. This was leading to the jobs running either until the
application stops or until the timeout was consumed.