cmake: enable warnings for testing code
Not having warnings makes it more likely to have buggy tests. A test
asserting a behavior incorrectly is no different from the behavior being
wrong in the first place.

...

The actual motivation is that I spent an evening debugging a "broken
promise" exception because of the missing `return` statement in one of
the branches, which the compiler didn't catch. I don't want that to
happen again.

Some checks are explicitly disabled for now to avoid having to refactor
too much code in one go. After the current change is merged we can work
on improving the code gradually.