transform/manager: don't move during erase
Since we can break out of the loop, it's possible that we move an entry
out of the btree_set secondary index and then that breaks the container,
causing assertion errors later on with the invalid index.

I wrote a unit test that caught this issue, then had to dump the index
multiple times to find what was going on.

Signed-off-by: Tyler Rockwood <rockwood@redpanda.com>