r/tests: made reconfiguration test more predictable
The `reconfiguration` test might failed when executed with leader acks.
In slow debug build modes and with tight intervals it might happend that
the leader stepped down after returning from replicate but before
actually replicating batches to the followers. If new leader is elected
in this case the log may be truncated and there may be no batches in the
log, making the test to fail.

Fixes: #14425

Signed-off-by: Michal Maslanka <michal@redpanda.com>