r/tests: don't loop forever
I have evidence that we're looping forever here because there is no exit
condition. I picked an arbitrary amount of time, so I'm happy to tweak it
or make it iteration based instead. We just need some exit condition.

Also clamp the reader so we don't try and read only data that isn't
replicated yet.

Signed-off-by: Tyler Rockwood <rockwood@redpanda.com>