resource_mgmt/memory_groups: fix memory groups on startup
The cluster config values can change, but in reality we only care about
the values on startup since they aren't read after the fact, so freeze
them on startup. We don't strictly enforce that initialize is called
however, as tests don't call this function and that would cause a lot
of churn for the tests.

Signed-off-by: Tyler Rockwood <rockwood@redpanda.com>