cloud_storage: Add validation metadata to the STM
Add new field to the 'segment' record that indicates that the segment
was validated. If the field is set to true this means that the
'ntp_archiver' invoked 'safe_segment_meta_to_add' method of the
'partition_manifest' to check the segment. The default value is 'false'
which means that the segment wasn't validated. This means that when we
will be replaying the log that contains version-0 records they won't be
validated inside the STM.

The ntp_archiver set's the flag to 'true' by defalult and if the checks
are disabled it set's the flag to 'false'.  If any inconsistencies are
added with disabled checks they will be applied to the STM.