storage: adjacent merge compact after windowed compaction
This is a stopgap solution to ensure the number of segments reduces
continually through the lifetime of a partition. Many tests make this
assumption, and the current approach of just rewriting individual
segments doesn't cut it.

With this commit, we will run merge compaction after windowed compaction
if windowed compaction didn't run. This condition isn't necessary but is
another expectation of many tests: that running housekeeping for the
first time doesn't reduce the number of segments.

Further work will be required to fix the solution in a way that doesn't
cause as much write amplification (e.g. by compacting into a single
segment per term).