audit_log_test: Mark test_audit_log_metrics ok_to_fail
mitigates: https://github.com/redpanda-data/redpanda/issues/14565

Tests in this file use the existence of the audit topic as a proxy
for full initialization of the audit manager. This is fine for the
general case (interacting with the audit log), but the client probe
is created after the topic, so we have a race in the test.

The fix will be straightforward, but I have a couple of questions
for Rob. Marking this ok_to_fail to avoid a bunch of noise on various
builds.