tests/topic_recovery_tests: archiver_term added to replaced segment key
cloud storage segments encode the archiver term in the object key. this
change is not reflected in the json translation of the binary partition
manifest, that generates keys without an archiver term.

this means that we can have these segment in s3,

1. 9452-10253-881631-1-v1.log.1
2. 9452-10255-881909-1-v1.log.2

have 1 listed in the replaced section of the binary partition manifest,
but the json version will list it as 9452-1-v1.log. this confuses the
TopicRecoveryTest.test_time_based_retention, that will consider both as
replaced and not count them up.

this pr fixes the test by mangling the archiver_term in the segment key

(cherry picked from commit 45a9ce1bcb8088d8fd0a622c0f3bef99bbded3be)