cluster: add unsafe reset from cloud for ntp manifest
`partition` gains a new function which resets the partition's manifest
to the manifest uploaded to cloud storage. If the downloaded manifest
has its last offset below the start offset of the local log, we refuse
the change since it would lead to data loss.

I also refactored the manifest reset methods a bit to share code and
added some comments to the method definitions.