s/a/schemas: Added cloud profile
The OCSF server reports that the use of the "api" object in the "API
Event" class can only be present if the "cloud" profile is selected.
The use of the "cloud" profile requires the use of the "cloud" object.
I believe this is a bug in their schema, however to satisfy the OCSF
server schema checker, we are introducing the "cloud" profile and
"cloud" object.

Signed-off-by: Michael Boquard <michael@redpanda.com>