rptest: Fix usage test incorrect assertion
- The usage_test was incorrectly asserting on a comparison of the values
of the bytes_in_cloud_storage field.

- Observed failures of this test saw the failure occur when comparing an
open window to a closed window. In this case the test will use le other
then eq to compare the fields, however this field is particularly
special as it is supposed to be set to -1 if the queried broker is not
the contoller leader. This test triggers a restart so it is common that
this value may change.

- The fix is to only compare these fields if the windows are closed,
that way the value is fixed in time across restarts.

- Fixes: #13448

(cherry picked from commit f1eaa00e8e13ffe93e2e3da66ae904396ad29de7)