ducktape/audit_log_test: Fix racy metrics test
The test harness uses the presence of the audit topic as a proxy
for audit log initialization. The audit client probe, however, is
constructed _after_ topic creation. For this test, we can just wait
for each node to export the client probe metric on the internal
endpoint before fetching and asserting on all expected metrics. It's
unlikely that we'll ever timeout here unless something else is broken
or changed.

Fixes #14565

Signed-off-by: Oren Leiman <oren.leiman@redpanda.com>