cluster: add force flag to reset from cloud
When resetting the manifest from cloud we check that the last offset in
the manifest is not below the max collectible offset of the partition as
a safety measure. In some circumstances, this check may be too strict:
read replicas and topics where remote write is turned off, will have
their collectible offset pegged at `offset::max`.

This commit adds a force flag to the admin request which can be used to
deal with such cases. The flag is `false` by default.