cluster: fix race between leader move and unsafe reset
Before transferring leadership, we inform the archiver via
`prepare_transfer_leadership`. Similarly, when the leadership transfer
is complete we call `complete_transfer_leadership`.

Previously, the code assumed that the archiver outlives the leadership
transfer. That's not necessarily the case. We destroy the archiver when
changing the topic configs and when performing a rest of the manifest
from the cloud.

The fix is to check if the archiver is alive after the leadership
transfer completed and call the respective method only in that case.

Why is this correct? `complete_transfer_leadership` unpauses the
archiver. If the archiver is not present at the point when we want to
unpause, it will be created later in its default unpaused state.