s/audit: Reject audit messages when misconfigured
- This change removes the conditional code that could have set up a non
ephemeral audit client.

- To identify an audit client as an epehemeral client, authn must be
enabled, to know if this is the case the client will act on reciept of
the illegal_sasl_state error code. Upon observation of this two things
will occur. Firstly, warning messages will arise to let the user know
that auditing has been misconfigured, and secondly a boolean will be set
that lets the audit_log_manager know this state has been entered. When
true, the audit_log_manager will reject future audit messages until the
issue has been resolved. This is to prevent the queues from filling up
and intentionally bring the issue to the users attention.