tests: introduce seastar task queue drain function
This drain function is the pattern used in transform_manager_test
hacking around idle_cpu_handler to tell us when the seastar queue is
empty.

There is already a function here to do this, but it crudely sleeps to
get around debug build task reordering, while this approach is more
robust and does not require the sleep hack.

Ideally `drain_task_queue` can replace all usages of `flush_tasks` but
I'll save that for another day.

Signed-off-by: Tyler Rockwood <rockwood@redpanda.com>