io: add storage provider
The storage provider gives access to storage for the io subsystem. There are two
implementations, one that passes directly to Seastar I/O's subsystem and is
intended for production cases. The second is an in-memory implementation that is
useful in testing for speed and for enabling parameter exploration (e.g.
alignments) that would not work on a real file system.

In addition to testing, the provider abstraction is a centralized
location to implementation fault injection, tracking open-files, and
allowing users (code) of I/O to be easier to identify.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>