cloud_storage: Track hydrations in remote segment
In remote segment, when we start hydration of a segment or chunk, we
increment the active hydration counter. This is decremented when the
hydration finishes (successful or failed). This also includes the time
taken to acquire client etc so the time during which a hydration is
active is slightly longer than the actual download time span.