wasm: refactor common Wasm execute
Wasmtime executes WebAssembly on a seperate stack, and the computation
running there is represented by a `wasmtime_call_future_t`. We call
`poll` to execute WebAssembly for some time and then yield.

We poll a WebAssembly function in two places:
1. When we create a WebAssembly VM the standard allows for "setup" code
   to run via a start function. None of the languages we support today
   use this functionality, but Wasmtime requires its usage anyways.
2. When we run `_start`, which runs the user's provided `main` function
   as defined in the wasi standard. This is a "long" lived function that
   runs as long as the VM is in our ABI contract.

Pull where we do this into a single place so that we can add timing only
once.

Signed-off-by: Tyler Rockwood <rockwood@redpanda.com>