cluster: introduce archiver mutex to partition
The partition may interact with the archiver at various points in its
lifetime:
* when it gets started/stopped
* when leadership moves
* when the topic configs are updated
* when the manifest needs to be reset from the cloud

These operations may race and lead to all sorts of issues like double
starting or double stopping the archiver. In order to avoid such cases,
a mutex is introduced to protect the archiver in this commit. It should
be grabbed every time when the partition needs to start and/or stop the
archiver.