transform: validate reader offset
There are cases when we want to prevent transforms from reading outside
the valid range. In order to do this, we start reading from the max of
requested offset and effective_start of the log.

The fetch handler validates the the fetch offset is >= effective_start,
so this should be similar to what we're doing there.

Signed-off-by: Tyler Rockwood <rockwood@redpanda.com>