tm_stm: add tx to cache on create
A test using transaction would randomly fail with
`INVALID_PRODUCER_ID_MAPPING` if the transaction coordinator leader
would change between the producer issued `InitProducerId` and
`AddPartitionsToTxn`. Since [Fixes Unknown Server Errors (USE) in
transactions][1], this should not happen as we cache transaction
state and fetch it from all nodes. However, it appears we forgot to
cache the state after the `InitProducerId` request. This PR fixes that
and updates a test that catches the problem and validates the fix.

Fixes #13736

[1]: https://github.com/redpanda-data/redpanda/pull/7591