rp_storage_tool: decode spillover manifets entries
Previously, rp_storage_tool did not decode the list of spillover
manifests encoded as a column store in
`partition_manifest::_spillover_manifests`. This was problematic for a
few reasons:
1. no way of inspecting the list of spillover manifests given a binary
   encoded partition manifest
2. tests that download the manifest, change it, and then perform a
   reset would lose the list of spillover manifests after the reset

This patch brings the rp_storage_tool capabilities largely in line with
the tip of Redpanda dev. The `metadata_size_hint` is ignored from the
segment metas, but that's likely fine for now.