rptest: simplify test_reset_spillover test
Now that rp_storage_tool decodes the list of spillover manifests from
the stm manifest, this test can be simplified. Previously, it had to
merge all segments from all spillover manifests into the STM manifset,
but now it can mutate the list of spillover manifests and re-upload to
achieve the same mutation.