cloud_storage: filter missing spills below archive start
Previously, we did not filter missing spillover manifests based on the
start offset of the archive. We simply checked if there exists a
matching spillover manifest in the manifest. This is problematic because
we remove the spillover manifest from the in-memory list after garbage
collection. Therefore, the scrubber can race with it and produce a false
positive.

To fix this, discard any missing spillover manifest anomaly that lies
below the `archive_start_offset`.