transform/processor: lag metrics
Create metrics for how far behind transforms are. This is important to
know if a user needs to scale up or improve the performance of their
transforms.

We aggregate the lag per transform, if a user needs to know the exact
lag for a single partition, `rpk transform list` can be used.

Signed-off-by: Tyler Rockwood <rockwood@redpanda.com>