doc: update style guide from discussion
This style guide was not updated in any meaningful way for >4 years. We
had started a github discussion, but that is too unstructured. I've
copied over parts of that discussion and will close the github
discussion. Making this be committed into the tree will force consensus
in the context of PRs.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>