tests/go: introduce transform-verifier
Introduce a small go program to validate transforms. kgo-verifier isn't
capable of testing transforms in the manner that is required.

The program uses a similar communication mechanism to kgo-verifier via a
small HTTP server that runs and reports the status. We also ensure
graceful remote shutdown so that ducktape can query the latest status
before the process exits stage left.

Signed-off-by: Tyler Rockwood <rockwood@redpanda.com>